+++ src/main/resources/templates/pollingunits.html
@@ -11,48 +11,21 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
         
     <link rel="stylesheet" type="text/css" th:href="@{/css/style.css}"/>
-    <!-- Global site tag (gtag.js) - Google Analytics -->
-    <script async src="https://www.googletagmanager.com/gtag/js?id=UA-141725675-1"></script>
-    <script type="text/javascript" th:src="@{/js/googleanalytics.js}"></script>
-	</head>
-    <!--Start of Tawk.to Script-->
-    <script type="text/javascript" th:src="@{/js/tawkto.js}"></script>
-    <!--End of Tawk.to Script-->
 	<body>
     <h2>Polling Units</h2>
-    <div>
-    <table class="striped responsive-table">
-        <tr>
-            <th>#</th>
-            <th><a th:href="@{?sortBy=fullCode&sortDirection=ASC}">Code</th>
-            <th><a th:href="@{?sortBy=description&sortDirection=ASC}">Description</th>
-            <th><a th:href="@{?sortBy=longitude&sortDirection=ASC}">Longitude</th>
-            <th><a th:href="@{?sortBy=latitude&sortDirection=ASC}">Latitude</th>
-        </tr>
-        <tr th:each="pollingUnit, i: ${form.pollingUnitsPage.content}">
-            <td th:text="${(form.pollingUnitsPage.number*form.pollingUnitsPage.size)+i.index+1}"></td>
-            <td style="white-space: nowrap;" th:text="${pollingUnit.fullCode}"></td>
-            <td th:text="${pollingUnit.description}"></td>
-            <td style="white-space: nowrap;" th:text="${pollingUnit.longitude}"></td>
-            <td style="white-space: nowrap;" th:text="${pollingUnit.latitude}"></td>
-        </tr>
-    </table>
-    </div>
+        <div id="react">
+            <!-- Global site tag (gtag.js) - Google Analytics -->
+            <script async src="https://www.googletagmanager.com/gtag/js?id=UA-141725675-1"></script>
+            <script type="text/javascript" th:src="@{/js/googleanalytics.js}"></script>
+            </head>
+            <!--Start of Tawk.to Script-->
+            <script type="text/javascript" th:src="@{/js/tawkto.js}"></script>
+            <!--End of Tawk.to Script-->
+            <!--JavaScript at end of body for optimized loading-->
+            <script type="text/javascript" th:src="@{/js/materialize.min.js}"></script>
 
-    <div class="pagination-div">
-        <a th:href="@{/pollingunits(page=0,size=${form.pollingUnitsPage.size})}">First</a>
-        <span th:if="${form.pollingUnitsPage.hasPrevious()}">
-            <a th:href="@{/pollingunits(page=${form.pollingUnitsPage.number-1},size=${form.pollingUnitsPage.size})}">Previous</a>
-        </span>
-        Page<span th:text="${form.pollingUnitsPage.number+1}"></span> of <span th:text="${form.pollingUnitsPage.totalPages}"></span>
-        <span th:if="${form.pollingUnitsPage.hasNext()}">
-            <a th:href="@{/pollingunits(page=${form.pollingUnitsPage.number+1},size=${form.pollingUnitsPage.size})}">Next</a>
-        </span>
-        <a th:href="@{/pollingunits(page=${form.pollingUnitsPage.totalPages-1},size=${form.pollingUnitsPage.size})}">Last</a>
-    </div>
-
-    <!--JavaScript at end of body for optimized loading-->
-    <script type="text/javascript" th:src="@{/js/materialize.min.js}"></script>
-
+            <!-- Load our React component. -->
+            <script th:src="@{/built/bundle.js}"></script>
+        </div>
     </body>
     </html>
